requirements.txt
bottle==0.12.16
suds-py3==1.3.3.0
